Few Changes
For the link fix changed the url from being 1_2 to being 0_2 in the html
file.

Changed the hyper link on "zipped" in the Rmd/pdf file.

To fix the swirl lessons I modified the .yaml files.  There was a case
in two of the files (Missing Values and Sequences of Numbers) where
ANY_of_exprs was written, but any_of_exprs is the correct form.