Subsetting Vectors Fixes
Originally there was an error revolving around the use of the yaml.load
function when trying to run the subsetting vectors lessons.  I read this
could sometimes be caused by unexpected whitespace, so i copied the
formatting of previous files and applied it to this one and it ran for
me.

Notifying instructor at end of lesson did not work at first, because the
notify() definition was mixing from the customTests file, so copied it
from a previous file and added it.  It seemed to work.